Restructure hook order for tabs
