+++ src/containers/Review/ReviewWrapper.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import { Route, Switch } from 'react-router'
-import { Message } from 'semantic-ui-react'
+import { Container, Message, Tab, Label, Icon, Header } from 'semantic-ui-react'
 import { Loading, NoMatch } from '../../components'
 import { useUserState } from '../../contexts/UserState'
+import useGetApplicationStructure from '../../utils/hooks/useGetApplicationStructure'
 import useGetReviewInfo from '../../utils/hooks/useGetReviewInfo'
 import { useRouter } from '../../utils/hooks/useRouter'
 import usePageTitle from '../../utils/hooks/usePageTitle'
@@ -20,37 +21,77 @@ const ReviewWrapper: React.FC<ReviewWrapperProps> = ({ structure }) => {
   const {
     match: { path },
   } = useRouter()
-  const {
-    userState: { currentUser },
-  } = useUserState()
+  const { error, fullStructure } = useGetApplicationStructure({
+    structure,
+    firstRunValidation: false,
+    shouldCalculateProgress: false,
+  })
 
   usePageTitle(strings.PAGE_TITLE_REVIEW.replace('%1', structure.info.serial))
 
-  // I think we need an option to selecte review assgnments where
-  // userId is reviewerId or assignerId or both or not match it at all (just by row level permission restrictions)
-  const { error, loading, assignments } = useGetReviewInfo({
-    applicationId: structure.info.id,
-    userId: currentUser?.userId as number,
-  })
+  if (error) return <Message error title={strings.ERROR_GENERIC} list={[error]} />
 
-  if (error) return <Message error header={strings.ERROR_REVIEW_PAGE} list={[error]} />
+  if (!fullStructure) return <Loading />
 
-  if (loading) return <Loading />
+  const {
+    info: { template, org, name },
+  } = fullStructure
 
-  if (!assignments || assignments.length === 0) return <NoMatch />
+  const tabPanes = [
+    { menuItem: 'Overview', render: () => <Tab.Pane>Overview</Tab.Pane> },
+    {
+      menuItem: 'Assignment',
+      render: () => (
+        <Tab.Pane>
+          <AssignmentWrapper {...{ structure }} />
+        </Tab.Pane>
+      ),
+    },
+  ]
 
   return (
-    <Switch>
-      <Route exact path={path}>
-        <AssignmentWrapper {...{ assignments, structure }} />
-      </Route>
-      <Route exact path={`${path}/:reviewId`}>
-        <ReviewPageWrapper {...{ structure, reviewAssignments: assignments }} />
-      </Route>
-      <Route>
-        <NoMatch />
-      </Route>
-    </Switch>
+    <Container id="review-area">
+      <Switch>
+        <Route exact path={path}>
+          <ReviewHomeHeader
+            templateCode={template.code}
+            applicationName={name}
+            orgName={org?.name as string}
+          />
+          <Tab panes={tabPanes} />
+        </Route>
+        <Route exact path={`${path}/:reviewId`}>
+          <ReviewPageWrapper {...{ structure }} />
+        </Route>
+        <Route>
+          <NoMatch />
+        </Route>
+      </Switch>
+    </Container>
+  )
+}
+
+interface ReviewHomeProps {
+  templateCode: string
+  applicationName: string
+  orgName?: string
+}
+
+const ReviewHomeHeader: React.FC<ReviewHomeProps> = ({
+  templateCode,
+  applicationName,
+  orgName,
+}) => {
+  const { push } = useRouter()
+  return (
+    <div id="review-home-header">
+      <Label
+        className="simple-label clickable"
+        onClick={() => push(`/applications?type=${templateCode}`)}
+        icon={<Icon name="chevron left" className="dark-grey" />}
+      />
+      <Header as="h3" content={applicationName} subheader={<Header as="h5" content={orgName} />} />
+    </div>
   )
 }
